# dbwrapper: Bump LevelDB max file size to 128 MiB to avoid system slowdown from high disk cache flush rate