@willcl-ark 2024-05-04
Might partially address https://github.com/bitcoin/bitcoin/issues/29662