@tdb3 2024-05-05
> > Are there any drawbacks to this?
> 
> I didn't notice any.
> 
> It's worth mentioning that the total amount of data stored in this database is at least two orders of magnitude higher than even 128 MiB file size.

It would be great if there are few/no drawbacks.  Do you mind sharing the methods used so far to test this?  It would be great to have some data for comparison.

Other questions that come to mind (thinking out loud before I dig deeper or perform testing):
- Does the change from 2MB to 128MB have any impact on consistent or transient RAM usage (i.e. for resource-constrained nodes)?
- Is the file size (or an option to use the legacy smaller file size) something we would want to expose in bitcoin.conf (e.g. as a debug option)?