@andrewtoth 2024-05-05
> Might partially address https://github.com/bitcoin/bitcoin/issues/29662

That issue is complaining about long compaction times. From https://github.com/bitcoin/bitcoin/blob/master/src/leveldb/include/leveldb/options.h#L111-L112:

> The downside will be longer compactions and hence longer latency/performance hiccups.

 it seems this change would make compaction times longer, so would exacerbate that issue?