@maciejsszmigiero 2024-05-05
> Do you mind sharing the methods used so far to test this? 

I am simply watching the disk cache flush rate in `iostat(1)`.
In addition to that, the difference in the system interactivity  is also pretty apparent.

> Does the change from 2MB to 128MB have any impact on consistent or transient RAM usage (i.e. for resource-constrained nodes)?

Did not observe any such effect, the RAM usage of the Bitcoin process seems to vary within roughly the same bounds when syncing with the Bitcoin network with our without this change.

> Is the file size (or an option to use the legacy smaller file size) something we would want to expose in bitcoin.conf (e.g. as a debug option)?

Maybe, but I don't know whether it makes sense to expose additional tuning option with respect to, for example, maintenance impact.

> > The downside will be longer compactions and hence longer latency/performance hiccups.

> it seems this change would make compaction times longer, so would exacerbate that issue?

For me, the biggest performance impact of compaction is from disk cache flushes this operation generates.
This patch significantly reduces such flush rate and so should make compaction less painful.