@andrewtoth 2024-05-06
Benchmarked IBD with an SSD to block 800k, `dbcache=450`, `prune=0` with a local node serving the blocks. This branch is 27% (!) faster than master :rocket: 


```
commit 7f15e71f7e762645dbd1ea5eba9ecc6f9ad60236 (branch)
  Time (mean ± σ):     14711.490 s ± 225.376 s    [User: 19465.517 s, System: 1147.712 s]
  Range (min … max):   14552.125 s … 14870.854 s    2 runs
  
 commit  eb0bdbdd753bca97120247b921fd29d606fea6e9 (master)
  Time (mean ± σ):     20274.276 s ± 106.042 s    [User: 21762.310 s, System: 4546.936 s]
  Range (min … max):   20199.293 s … 20349.259 s    2 runs
```

> This patch significantly reduces such flush rate and so should make compaction less painful.

From what I understand, this patch reduces the frequency of flushes, but they will take longer when they do occur. This is great for IBD, but for https://github.com/bitcoin/bitcoin/issues/29662 the issue is an unavoidable compaction at startup. The compaction could potentially take longer with this patch.