@laanwj 2024-05-06
> This branch is 27% (!) faster than master

That's impressive!

> From what I understand, this patch reduces the frequency of flushes

Not only the frequency of flushes; another potential advantage here is that leveldb will spend less time open()ing and close()ing files to maintain its allowed number of open files (eg the `fd_limiter` stuff).