@andrewtoth 2024-05-08
@maciejsszmigiero why not increase `max_file_size` even more? Are there drawbacks to increasing it to 256 MB or 1 GB?